+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 ## [Unreleased]
+### Fixed
+- Coinbase parser: fixed negative values for Advanced Trade Sells.
+### Added
+- Coinbase parser: added "Cash to Savings" and "Interest payout" transaction types.
 
 ## Version [0.6.0] (2025-11-05)
 Important:-